Start work on vim gui object.
Will be class that contains other widgets. It will just process
keypresses before handing them off to anything inside of it.

This is just a small work in progress, currently, to see what's
possible.

I've created a simple status line to display the current mode and
created a signal so that the mode displayed is changed when the mode is
changed.

 Changes to be committed:
	new file:   vim_wrapper.py