Implement working controls for all command_map commands.
Work-in-progress. Still need to figure out how to handle the 'esc' key,
since the wrapper depends on it when in INSERT mode, even though all
keys should be passed to the underlying widget while in INSERT mode.

Also need to think about non-standard commands, like half-page up or
half-page down. Am I going to have to change/extend widgets directly for
this functionality?

 Changes to be committed:
	modified:   vim_wrapper.py